move sound logic to stream hook, add fadeout logic
